fix(fbemitter): fixed the error in the event emitter
